refactor : 간단한 코드 포맷 변경 및 파일 구조 변경
