+++ src/main/kotlin/com/teamsparta/mini5foodfeed/common/exception/dto/ErrorResponse.kt
@@ -1,0 +1,5 @@
+package com.teamsparta.mini5foodfeed.common.exception.dto
+
+data class ErrorResponse(
+    val message: String?
+)
