+++ src/main/kotlin/com/teamsparta/mini5foodfeed/domain/comment/repository/CommentRepository.kt
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CommentRepository: JpaRepository<Comment, Long> {
+interface CommentRepository : JpaRepository<Comment, Long> {
 
     @Query("SELECT c FROM Comment c LEFT JOIN FETCH c.feed f WHERE f.id = :feedId ORDER BY c.createdAt DESC")
     fun findTop5ByFeedIdOrderByCreatedAtDesc(feedId: Long?, pageable: Pageable): List<Comment>
