+++ src/main/kotlin/com/teamsparta/mini5foodfeed/domain/user/exception/UserIdNotFoundException.kt
@@ -1,3 +1,3 @@
 package com.teamsparta.mini5foodfeed.domain.user.exception
 
-data class UserIdNotFoundException(val userId: String): RuntimeException("Not Found UserId: $userId")
+data class UserIdNotFoundException(val userId: String) : RuntimeException("Not Found UserId: $userId")
