+++ src/main/kotlin/com/teamsparta/mini5foodfeed/domain/user/exception/userIdIllegalStateException.kt
@@ -1,3 +1,3 @@
 package com.teamsparta.mini5foodfeed.domain.user.exception
 
-data class UserIdIllegalStateException(val userId:String):RuntimeException("already use in username:$userId")
+data class UserIdIllegalStateException(val userId: String) : RuntimeException("already use in username:$userId")
