--- src/main/kotlin/com/teamsparta/mini5foodfeed/exception/NotFoundException.kt
@@ -1,3 +1,0 @@
-package com.teamsparta.mini5foodfeed.exception
-
-data class NotFoundException(val id:Long):RuntimeException(" Not Found with give id:$id")
