--- src/main/kotlin/com/teamsparta/mini5foodfeed/exception/dto/ErrorResponse.kt
@@ -1,5 +1,0 @@
-package com.teamsparta.mini5foodfeed.exception.dto
-
-data class ErrorResponse(
-    val message: String?
-)
